Decouple vulkan stream definition into a separate file
